`torch.log_softmax(bf16_in)` returns bf16, and the per-element storage
precision loses the tail of the distribution: on Qwen2.5-1.5B-Instruct the
L-infinity gap against an fp32 softmax runs 0.008 (head) to 0.126 (deep tail)
on HF and 0.062 to 0.187 on MLX.

The MLX side conflicted: the branch predates the KV trie-to-slot rewrite, so
its `add_to_cache`/`_batch_logits_custom` no longer exist. Resolved to the
current `_forward`, casting the slot's logits to fp32 before the logsumexp.

vLLM and SGL already do the softmax in fp32 internally; neither is touched.

+++ genlm/backend/llm/mlx.py
@@ -289,7 +289,7 @@ else:
             """Next-token log-probs under one adapter, ``[len(prompts), vocab]``."""
             with wired_limit(self.mlx_lm_model, [self.generation_stream]):
                 self.adapters.select(lora_name)
-                logits = self.slots[lora_name].logits(prompts)
+                logits = self.slots[lora_name].logits(prompts).astype(mx.float32)
                 logprobs = logits - mx.logsumexp(logits, axis=-1, keepdims=True)
                 mx.eval(logprobs)
             return _to_torch(logprobs)
